decouple watcher from the workspace. move the watcher to be part of the workspace extension, remove the compiler dependency, add slot onComponentChange that gets triggered once the watcher recongizes changes. in the compiler, register to this slot
