+++ src/extensions/workspace/on-component-change.ts
@@ -1,0 +1,5 @@
+import { Component } from '../component';
+
+export type OnComponentChangeOptions = { noCache?: boolean; verbose?: boolean };
+
+export type OnComponentChange = (component: Component, options: OnComponentChangeOptions) => Promise<any>;
