--- src/extensions/workspace/watch/index.ts
@@ -1,1 +1,0 @@
-export { WatcherExtension } from './watch.extension';
